updated args.folderlocation to trimmed_dir
The plasmidspades part of the assembly_pipeline script referenced args.folderlocation, which was an artifact of a previous version. I changed this to trimmed_dir, which is a variable that leads to a directory containing trimmed reads.

We should discuss whether to hardcode the locations of the trimmed data and QC reports or make them arguments. For now, I'm just hardcoding them.